backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
groq==0.9.0
faster-whisper==1.0.3
azure-cognitiveservices-speech==1.40.0
sqlalchemy==2.0.30
requests==2.32.3
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
pydantic-settings==2.3.4
pydub==0.25.1
